+++ CHANGELOG.md
@@ -38,6 +38,7 @@ _Not yet on NuGet..._
 * Scatter: Added support for `Scale` and `Offset` properties (#3835) @bukkideme
 * Axis Lines: Separated `LegendText` from `LabelText` so items may be configured separately
 * Heatmap: Exposed `CellWidth` and `CellHeight` as an alternative sizing strategy to setting `Extent` (#3869) @alexisvrignaud
+* ImageRect: New plot type that places an image inside a defined rectangle on the plot (#3870) @sdpenner
 
 ## ScottPlot 5.0.34
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-05-05_